tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-slack-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: masonry-slack parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="masonry-slack supports the full grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("masonry-slack", "normal");

test_valid_value("masonry-slack", "0", "0px");
test_valid_value("masonry-slack", "1px");
test_valid_value("masonry-slack", "calc(2em + 3ex)");
test_valid_value("masonry-slack", "4%");
test_valid_value("masonry-slack", "5vmin");
</script>
</body>
</html>
